requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
google-generativeai==0.3.0
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
